Added detection for touch devices.  However, detection is only possible after the user touches something.
Any button with a tooltip requires two taps the first time it is used.  I used this detection to disable tooltips on the enlarge buttons.  Unable to disable tooltip on index since it loads before the user can interact with the page.  We may want to try something like this, which claims the ability to disable on touch devices:  http://iamceege.github.io/tooltipster/