SSBART - Tab Order
Ensure that the reading order of content is logical	The tab order is: next, back, index button, and glossary button. However, the reading order is reversed: index button, glossary button, back, and next.	Rearrange the buttons in the DOM so they are similar to the tab order instead of reverse like it is now.